dashboard/requirements.txt
streamlit>=1.30.0
plotly>=5.18.0
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0
pyyaml>=6.0
requests>=2.28.0
holidays>=0.40
scikit-learn>=1.3.0
